requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# Data Science and Machine Learning
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.3.0

# Deep Learning
tensorflow>=2.8.0
# torch>=1.12.0  # Alternative to TensorFlow
# torchvision>=0.13.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Image Processing
imageio>=2.9.0
skimage>=0.18.0

# Progress bars and utilities
tqdm>=4.62.0
joblib>=1.1.0

# Configuration
PyYAML>=6.0
configparser>=5.2.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0
